MatSim/requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.3
fonttools==4.61.1
geopandas==1.1.2
idna==3.11
ipfn==1.4.4
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
libpysal==4.14.1
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.1
packaging==25.0
pandas==2.3.3
pillow==12.1.0
platformdirs==4.5.1
pointpats==2.5.2
pyogrio==0.12.1
pyparsing==3.3.1
pyproj==3.7.2
pyreproj==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
shapely==2.1.2
six==1.17.0
soupsieve==2.8.1
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.6
